Move to a src layout and don't package tests
The community has been moving to putting the package in a src folder.
This has some advantages, for example we don't need to make a temp
folder to run the tests in since the module isn't importable from the
repository root without being installed. Also moving the tests to
a folder outside of the package to avoid distributing them. The tests
don't need to be there and aren't used outside of development. This
makes the tarballs and wheels a bit smaller, which is good.
