--- .coveragerc
@@ -1,6 +1,0 @@
-[run]
-branch = True
-omit =
-    */tests/*
-    */_version_generated.py
-    *verde/__init__.py
