+++ .flake8
@@ -16,13 +16,13 @@ exclude =
     .git,
     __pycache__,
     .ipynb_checkpoints,
-    verde/_version_generated.py,
+    src/verde/_version.py,
     doc/_build,
 per-file-ignores =
     # disable unused-imports errors on __init__.py
     __init__.py: F401
     # disable shadowing of builtin io module (this should be solved)
-    verde/io.py: A005
+    src/verde/io.py: A005
 
 # Configure flake8-rst-docstrings
 # -------------------------------
