+++ .gitignore
@@ -10,9 +10,6 @@ dist/
 .cache
 doc/_build
 doc/api/generated
-doc/gallery
-doc/tutorials
-doc/sample_data
 .ipynb_checkpoints
 *.eggs
 *.egg-info
@@ -20,5 +17,5 @@ MANIFEST
 dask-worker-space
 .coverage.*
 *.ipynb
-**/_version_generated.py
+**/_version.py
 sg_execution_times.rst
