+++ MANIFEST.in
@@ -1,13 +1,13 @@
 # Exclude these files from source distributions.
 # setuptools_scm includes everything else by default.
 prune .github
-prune data
-prune doc
 prune env
-prune paper
-prune tools
+prune doc
+prune test
 exclude .*.yml
 exclude .*rc
-exclude Makefile
 exclude .gitignore
+exclude CONTRIBUTING.md
+exclude CODE_OF_CONDUCT.md
 exclude environment.yml
+exclude Makefile
