+++ src/verde/__init__.py
@@ -38,3 +38,6 @@ from .spline import Spline, SplineCV
 from .trend import Trend
 from .utils import grid_to_table, make_xarray_grid, maxabs, variance_to_weights
 from .vector import Vector, VectorSpline2D
+
+# Append a leading "v" to the generated version by setuptools_scm
+__version__ = f"v{__version__}"
