+++ src/verde/base/least_squares.py
