+++ src/verde/blockreduce.py
