+++ src/verde/neighbors.py
