+++ src/verde/trend.py
