+++ src/verde/vector.py
